Move `SpliceStyle`, `MethodId`, `CallDisplayStyle` and `OptionalRangeInclusive` out of `query`
