+++ monument/lib/src/query.rs
@@ -1,11 +1,12 @@
 //! Instructions for Monument about what compositions should be generated.
 
-use std::ops::{Range, RangeInclusive};
+use std::ops::RangeInclusive;
 
-use bellframe::{music::Pattern, Bell, Mask, PlaceNot, RowBuf, Stage, Stroke};
+use bellframe::{music::Pattern, Mask, PlaceNot, RowBuf, Stage, Stroke};
 use serde::Deserialize;
 
 use crate::{
+    builder::{CallDisplayStyle, OptionalRangeInclusive, SpliceStyle},
     group::PartHeadGroup,
     utils::{PerPartLength, Score, TotalLength},
 };
@@ -75,29 +76,6 @@ impl Query {
 // METHODS //
 /////////////
 
-/// The different styles of spliced that can be generated
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash, Deserialize)]
-pub enum SpliceStyle {
-    /// Splices could happen at any lead label
-    #[serde(rename = "leads")]
-    LeadLabels,
-    /// Splices only happen at calls
-    #[serde(rename = "calls")]
-    Calls,
-}
-
-impl Default for SpliceStyle {
-    fn default() -> Self {
-        Self::LeadLabels
-    }
-}
-
-/// The unique identifier for a method in a [`Search`](crate::Search).
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct MethodId {
-    pub(crate) index: MethodIdx,
-}
-
 /// A `Method` used in a [`Query`].
 #[derive(Debug, Clone)]
 pub(crate) struct Method {
@@ -147,15 +125,6 @@ pub(crate) struct Call {
     pub(crate) place_notation: PlaceNot,
 
     pub(crate) weight: Score,
-}
-
-/// How the calls in a given composition should be displayed
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CallDisplayStyle {
-    /// Calls should be displayed as a count since the last course head
-    Positional,
-    /// Calls should be displayed based on the position of the provided 'observation' [`Bell`]
-    CallingPositions(Bell),
 }
 
 ///////////
@@ -211,48 +180,6 @@ impl Default for StrokeSet {
 ////////////////
 // MISC TYPES //
 ////////////////
-
-/// An inclusive range where each side is optionally bounded.
-///
-/// This is essentially a combination of [`RangeInclusive`](std::ops::RangeInclusive)
-/// (`min..=max`), [`RangeToInclusive`](std::ops::RangeToInclusive) (`..=max`),
-/// [`RangeFrom`](std::ops::RangeFrom) (`min..`) and [`RangeFull`](std::ops::RangeFull) (`..`).
-#[derive(Debug, Clone, Copy, Default)]
-pub struct OptionalRangeInclusive {
-    pub min: Option<usize>,
-    pub max: Option<usize>,
-}
-
-impl OptionalRangeInclusive {
-    /// An [`OptionalRangeInclusive`] which is unbounded at both ends.  Equivalent to
-    /// `OptionalRangeInclusive { min: None, max: None }`.
-    pub const OPEN: Self = Self {
-        min: None,
-        max: None,
-    };
-
-    /// Returns `true` if at least one of `min` or `max` is set
-    pub fn is_set(self) -> bool {
-        self.min.is_some() || self.max.is_some()
-    }
-
-    /// Applies [`Option::or`] to both `min` and `max`
-    pub fn or(self, other: Self) -> Self {
-        Self {
-            min: self.min.or(other.min),
-            max: self.max.or(other.max),
-        }
-    }
-
-    pub fn or_range(self, other: &Range<usize>) -> Range<usize> {
-        let min = self.min.unwrap_or(other.start);
-        let max = self
-            .max
-            .map(|x| x + 1) // +1 because `OptRange` is inclusive
-            .unwrap_or(other.end);
-        min..max
-    }
-}
 
 index_vec::define_index_type! { pub(crate) struct MethodIdx = usize; }
 index_vec::define_index_type! { pub(crate) struct CallIdx = usize; }
